Updated to balancing
Tried updating the balancing for 1/1 rolls to 20/20 rolls.  Ideally, I'd like the character to receive an even slope of 0gp for 1/1 to 13,000g for a 20/20.  But the reality was that either they received no gold for a 10/10 or 600,000g for a 20/20.  I ended up using a hard cap because I couldn't balance it right.